+++ out/compilers/dasmCompiler.js
@@ -125,18 +125,18 @@ class DasmCompiler extends compilerBase_1.CompilerBase {
                 this.CustomFolderOrPath = true;
             }
             else {
-                // dasm command (depends on OS)
-                let dasmCommand = "dasm.exe";
+                // dasm name (depends on OS)
+                let dasmName = "dasm.exe";
                 if (application.IsLinux) {
                     // Linux
-                    dasmCommand = "dasm.Linux.x86";
+                    dasmName = "dasm";
                 }
                 else if (application.IsMacOS) {
                     // MacOS
-                    dasmCommand = "dasm.Darwin.x86";
+                    dasmName = "dasm";
                 }
-                // Use the default
-                this.FolderOrPath = path.join(this.DefaultFolderOrPath, dasmCommand);
+                // Append path (based on architecture and emulator name)
+                this.FolderOrPath = path.join(this.DefaultFolderOrPath, application.OSPlatform, application.OSArch, dasmName);
             }
             // Compiler (other)
             this.Format = this.Configuration.get(`compiler.${this.Id}.format`, "3");