+++ CHANGELOG.md
@@ -4,9 +4,10 @@ The following enhancements and changes have been made to Atari Dev Studio:
 
 ## 0.3.1
 
-* Updated internal dev packages
+* Officially (finally!) included compiler and emulator packages for macOS (Mojave).
 * Added Stella 6.0.2 (macOS)
-* Offically tested compiler and emulator packages on macOS (Mojave)
+* Updated dasm to the latest 32 and 64 bit packages (Windows, Linux and macOS).
+* Updated internal dev packages 
 
 ## 0.3.0